Add back support for running on wasm
The PyO3 bindings had to be moved out into a separate implementation.

Also make application mutable in update.
