+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = ["run-wasm", "visula", "visula_derive", "visula_core", "visula_pyo3"]
 resolver = "2"
 
+
 [workspace.dependencies]
 wgpu = { version = "0.19", features = ["webgl"] }
 glam = { version = "0.24", features = ["bytemuck"] }
@@ -14,10 +15,10 @@ uuid = { version = "1.3.3", features = [
 ] }
 bytemuck = { version = "1.4", features = ["derive"] }
 log = "0.4"
-winit = { version = "0.29", default-features = false }
+winit = { version = "0.29" }
 egui = "0.26"
 egui-wgpu = { version = "0.26", features = ["winit"]  }
-egui-winit = { version = "0.26"  }
+egui-winit = { version = "0.26", default-features = false }
 pyo3 = { version = "0.21" }
 numpy = { version = "0.21", features = ["nalgebra"] }
 pollster = "0.3.0"
