+++ visula/examples/clay.rs
@@ -356,7 +356,7 @@ impl Simulation {
 
 impl visula::Simulation for Simulation {
     type Error = Error;
-    fn update(&mut self, application: &visula::Application) {
+    fn update(&mut self, application: &mut visula::Application) {
         let previous_particles = self.particles.clone();
         integrate(
             &mut self.particles,
