+++ visula/examples/neuronify.rs
@@ -697,7 +697,7 @@ impl Simulation {
 
 impl visula::Simulation for Simulation {
     type Error = Error;
-    fn update(&mut self, application: &visula::Application) {
+    fn update(&mut self, application: &mut visula::Application) {
         let Simulation {
             connection_tool,
             world,
